requirements.txt
tensorflow-cpu==2.6.4
fastapi
uvicorn
pillow
numpy
gunicorn
protobuf==3.20.*
python-multipart